refactor: utility based on DataConverter
